IBX-10122: Fix misspelled includedIcon filename
